+++ tests/test_httpx_async.py
@@ -48,9 +48,8 @@ async def test_url_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_url_matching_reusing_response(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(url="https://test_url")
+    httpx_mock.add_response(url="https://test_url", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -61,9 +60,8 @@ async def test_url_matching_reusing_response(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_url_query_string_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(url="https://test_url?a=1&b=2")
+    httpx_mock.add_response(url="https://test_url?a=1&b=2", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.post("https://test_url?a=1&b=2")
@@ -75,11 +73,9 @@ async def test_url_query_string_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(url="https://test_url")
+    httpx_mock.add_response(url="https://test_url", is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -92,11 +88,9 @@ async def test_url_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_query_string_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(url="https://test_url?a=1&a=2")
+    httpx_mock.add_response(url="https://test_url?a=1&a=2", is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -110,9 +104,8 @@ async def test_url_query_string_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_method_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(method="get")
+    httpx_mock.add_response(method="get", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -123,11 +116,9 @@ async def test_method_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_method_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(method="get")
+    httpx_mock.add_response(method="get", is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -140,9 +131,8 @@ async def test_method_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_reusing_one_response(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(url="https://test_url", content=b"test content")
+    httpx_mock.add_response(url="https://test_url", content=b"test content", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -171,11 +161,11 @@ async def test_response_with_html_string_body(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_stream_response_streaming(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         url="https://test_url",
         stream=pytest_httpx.IteratorStream([b"part 1", b"part 2"]),
+        is_reusable=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -201,11 +191,11 @@ async def test_stream_response_streaming(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_content_response_streaming(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         url="https://test_url",
         content=b"part 1 and 2",
+        is_reusable=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -229,11 +219,11 @@ async def test_content_response_streaming(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_text_response_streaming(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         url="https://test_url",
         text="part 1 and 2",
+        is_reusable=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -257,9 +247,8 @@ async def test_text_response_streaming(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_default_response_streaming(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         async with client.stream(method="GET", url="https://test_url") as response:
@@ -295,10 +284,9 @@ async def test_with_many_responses(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_with_many_reused_responses(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(url="https://test_url", content=b"test content 1")
-    httpx_mock.add_response(url="https://test_url", content=b"test content 2")
+    httpx_mock.add_response(url="https://test_url", content=b"test content 2", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -730,9 +718,8 @@ async def test_requests_retrieval(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_on_same_url(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(url="https://test_url")
+    httpx_mock.add_response(url="https://test_url", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url", headers={"X-TEST": "test header 1"})
@@ -745,9 +732,8 @@ async def test_requests_retrieval_on_same_url(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_request_retrieval_on_same_url(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url", headers={"X-TEST": "test header 1"})
@@ -758,9 +744,8 @@ async def test_request_retrieval_on_same_url(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_on_same_method(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url", headers={"X-TEST": "test header 1"})
@@ -773,9 +758,8 @@ async def test_requests_retrieval_on_same_method(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_request_retrieval_on_same_method(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url", headers={"X-TEST": "test header 1"})
@@ -786,9 +770,8 @@ async def test_request_retrieval_on_same_method(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url", headers={"X-TEST": "test header 1"})
@@ -803,9 +786,8 @@ async def test_requests_retrieval_on_same_url_and_method(httpx_mock: HTTPXMock) 
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_default_requests_retrieval(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.post("https://test_url", headers={"X-TEST": "test header 1"})
@@ -938,12 +920,11 @@ async def test_async_callback_returning_response(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_callback_executed_twice(httpx_mock: HTTPXMock) -> None:
     def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content"])
 
-    httpx_mock.add_callback(custom_response)
+    httpx_mock.add_callback(custom_response, is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -956,12 +937,11 @@ async def test_callback_executed_twice(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_async_callback_executed_twice(httpx_mock: HTTPXMock) -> None:
     async def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content"])
 
-    httpx_mock.add_callback(custom_response)
+    httpx_mock.add_callback(custom_response, is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -974,13 +954,12 @@ async def test_async_callback_executed_twice(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_callback_registered_after_response(httpx_mock: HTTPXMock) -> None:
     def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content2"])
 
     httpx_mock.add_response(json=["content1"])
-    httpx_mock.add_callback(custom_response)
+    httpx_mock.add_callback(custom_response, is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -998,13 +977,12 @@ async def test_callback_registered_after_response(httpx_mock: HTTPXMock) -> None
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_async_callback_registered_after_response(httpx_mock: HTTPXMock) -> None:
     async def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content2"])
 
     httpx_mock.add_response(json=["content1"])
-    httpx_mock.add_callback(custom_response)
+    httpx_mock.add_callback(custom_response, is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -1022,13 +1000,12 @@ async def test_async_callback_registered_after_response(httpx_mock: HTTPXMock) -
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_response_registered_after_callback(httpx_mock: HTTPXMock) -> None:
     def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content1"])
 
     httpx_mock.add_callback(custom_response)
-    httpx_mock.add_response(json=["content2"])
+    httpx_mock.add_response(json=["content2"], is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -1046,13 +1023,12 @@ async def test_response_registered_after_callback(httpx_mock: HTTPXMock) -> None
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_response_registered_after_async_callback(httpx_mock: HTTPXMock) -> None:
     async def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content1"])
 
     httpx_mock.add_callback(custom_response)
-    httpx_mock.add_response(json=["content2"])
+    httpx_mock.add_response(json=["content2"], is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -1070,12 +1046,11 @@ async def test_response_registered_after_async_callback(httpx_mock: HTTPXMock) -
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_callback_matching_method(httpx_mock: HTTPXMock) -> None:
     def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content"])
 
-    httpx_mock.add_callback(custom_response, method="GET")
+    httpx_mock.add_callback(custom_response, method="GET", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -1088,12 +1063,11 @@ async def test_callback_matching_method(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_async_callback_matching_method(httpx_mock: HTTPXMock) -> None:
     async def custom_response(request: httpx.Request) -> httpx.Response:
         return httpx.Response(status_code=200, json=["content"])
 
-    httpx_mock.add_callback(custom_response, method="GET")
+    httpx_mock.add_callback(custom_response, method="GET", is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         response = await client.get("https://test_url")
@@ -1111,14 +1085,13 @@ def test_request_retrieval_with_more_than_one(testdir: Testdir) -> None:
     """
     testdir.makepyfile(
         """
-        import pytest
         import httpx
+        import pytest
         
         
         @pytest.mark.asyncio
-        @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
         async def test_request_retrieval_with_more_than_one(httpx_mock):
-            httpx_mock.add_response()
+            httpx_mock.add_response(is_reusable=True)
         
             async with httpx.AsyncClient() as client:
                 await client.get("https://test_url", headers={"X-TEST": "test header 1"})
@@ -1131,7 +1104,7 @@ def test_request_retrieval_with_more_than_one(testdir: Testdir) -> None:
     result.assert_outcomes(failed=1)
     result.stdout.fnmatch_lines(
         [
-            "*AssertionError: More than one request (2) matched, use get_requests instead."
+            "*AssertionError: More than one request (2) matched, use get_requests instead or refine your filters."
         ]
     )
 
@@ -1160,13 +1133,11 @@ async def test_multi_value_headers_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_multi_value_headers_not_matching_single_value_issued(
     httpx_mock: HTTPXMock,
 ) -> None:
-    httpx_mock.add_response(match_headers={"my-custom-header": "value1"})
+    httpx_mock.add_response(match_headers={"my-custom-header": "value1"}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1185,13 +1156,11 @@ async def test_multi_value_headers_not_matching_single_value_issued(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_multi_value_headers_not_matching_multi_value_issued(
     httpx_mock: HTTPXMock,
 ) -> None:
-    httpx_mock.add_response(match_headers={"my-custom-header": "value1, value2"})
+    httpx_mock.add_response(match_headers={"my-custom-header": "value1, value2"}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1210,12 +1179,10 @@ async def test_multi_value_headers_not_matching_multi_value_issued(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_matching_respect_case(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
-        match_headers={"user-agent": f"python-httpx/{httpx.__version__}"}
+        match_headers={"user-agent": f"python-httpx/{httpx.__version__}"}, is_optional=True
     )
 
     async with httpx.AsyncClient() as client:
@@ -1229,16 +1196,15 @@ async def test_headers_matching_respect_case(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_not_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         match_headers={
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url2",
             "Host2": "test_url",
-        }
+        },
+        is_optional=True
     )
 
     async with httpx.AsyncClient() as client:
@@ -1252,9 +1218,7 @@ async def test_headers_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_not_matching_upper_case_headers_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1262,6 +1226,7 @@ async def test_url_not_matching_upper_case_headers_matching(
         method="GET",
         url="https://test_url?q=b",
         match_headers={"MyHeader": "Something"},
+        is_optional=True,
     )
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1292,11 +1257,9 @@ async def test_proxy_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_proxy_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(proxy_url="http://my_test_proxy")
+    httpx_mock.add_response(proxy_url="http://my_test_proxy", is_optional=True)
 
     async with httpx.AsyncClient(proxy="http://my_test_proxy") as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1309,11 +1272,9 @@ async def test_proxy_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_proxy_not_existing(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(proxy_url="http://my_test_proxy")
+    httpx_mock.add_response(proxy_url="http://my_test_proxy", is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1326,9 +1287,8 @@ async def test_proxy_not_existing(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_content_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.post("https://test_url", content=b"This is the body")
@@ -1339,9 +1299,8 @@ async def test_requests_retrieval_content_matching(httpx_mock: HTTPXMock) -> Non
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_json_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.post("https://test_url", json=["my_str"])
@@ -1352,9 +1311,8 @@ async def test_requests_retrieval_json_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_proxy_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient(
         mounts={
@@ -1374,9 +1332,8 @@ async def test_requests_retrieval_proxy_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_request_retrieval_proxy_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient(
         mounts={
@@ -1394,11 +1351,10 @@ async def test_request_retrieval_proxy_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_files_and_data_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.put(
@@ -1429,9 +1385,8 @@ async def test_requests_retrieval_files_and_data_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_request_retrieval_files_and_data_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.put(
@@ -1449,9 +1404,8 @@ async def test_request_retrieval_files_and_data_matching(httpx_mock: HTTPXMock) 
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_requests_retrieval_extensions_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url")
@@ -1470,9 +1424,8 @@ async def test_requests_retrieval_extensions_matching(httpx_mock: HTTPXMock) -> 
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_request_retrieval_extensions_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response()
+    httpx_mock.add_response(is_reusable=True)
 
     async with httpx.AsyncClient() as client:
         await client.get("https://test_url", timeout=httpx.Timeout(5, read=10))
@@ -1485,11 +1438,9 @@ async def test_request_retrieval_extensions_matching(httpx_mock: HTTPXMock) -> N
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_content_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(match_content=b"This is the body")
+    httpx_mock.add_response(match_content=b"This is the body", is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1520,11 +1471,9 @@ async def test_json_partial_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_json_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(match_json={"a": 1, "b": 2})
+    httpx_mock.add_response(match_json={"a": 1, "b": 2}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1537,13 +1486,12 @@ async def test_json_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_and_json_not_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         match_json={"a": 1, "b": 2},
-        match_headers={"foo": "bar"},
+        match_headers={"foo": "bar"}, 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1557,11 +1505,9 @@ async def test_headers_and_json_not_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_match_json_invalid_json(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(match_json={"a": 1, "b": 2})
+    httpx_mock.add_response(match_json={"a": 1, "b": 2}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -1586,16 +1532,15 @@ async def test_headers_and_content_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_not_matching_and_content_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         match_headers={
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url2",
         },
-        match_content=b"This is the body",
+        match_content=b"This is the body", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1609,16 +1554,15 @@ async def test_headers_not_matching_and_content_matching(httpx_mock: HTTPXMock) 
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_matching_and_content_not_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         match_headers={
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url",
         },
-        match_content=b"This is the body2",
+        match_content=b"This is the body2", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1632,16 +1576,15 @@ async def test_headers_matching_and_content_not_matching(httpx_mock: HTTPXMock) 
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_and_content_not_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         match_headers={
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url2",
         },
-        match_content=b"This is the body2",
+        match_content=b"This is the body2", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1668,9 +1611,7 @@ async def test_url_and_headers_and_content_matching(httpx_mock: HTTPXMock) -> No
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_not_matching_and_url_and_content_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1680,7 +1621,8 @@ async def test_headers_not_matching_and_url_and_content_matching(
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url2",
         },
-        match_content=b"This is the body",
+        match_content=b"This is the body", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1694,9 +1636,7 @@ async def test_headers_not_matching_and_url_and_content_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_and_headers_not_matching_and_content_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1706,7 +1646,8 @@ async def test_url_and_headers_not_matching_and_content_matching(
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url2",
         },
-        match_content=b"This is the body",
+        match_content=b"This is the body", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1720,9 +1661,7 @@ async def test_url_and_headers_not_matching_and_content_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_and_headers_matching_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1732,7 +1671,8 @@ async def test_url_and_headers_matching_and_content_not_matching(
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url",
         },
-        match_content=b"This is the body2",
+        match_content=b"This is the body2", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1746,9 +1686,7 @@ async def test_url_and_headers_matching_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_matching_and_url_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1758,7 +1696,8 @@ async def test_headers_matching_and_url_and_content_not_matching(
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url",
         },
-        match_content=b"This is the body2",
+        match_content=b"This is the body2", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1772,9 +1711,7 @@ async def test_headers_matching_and_url_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_matching_and_headers_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1784,7 +1721,8 @@ async def test_url_matching_and_headers_and_content_not_matching(
             "User-Agent": f"python-httpx/{httpx.__version__}",
             "Host": "test_url2",
         },
-        match_content=b"This is the body2",
+        match_content=b"This is the body2", 
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1798,9 +1736,7 @@ async def test_url_matching_and_headers_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_and_headers_and_content_not_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
         url="https://test_url2",
@@ -1809,6 +1745,7 @@ async def test_url_and_headers_and_content_not_matching(httpx_mock: HTTPXMock) -
             "Host": "test_url2",
         },
         match_content=b"This is the body2",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1838,9 +1775,7 @@ async def test_method_and_url_and_headers_and_content_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_headers_not_matching_and_method_and_url_and_content_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1852,6 +1787,7 @@ async def test_headers_not_matching_and_method_and_url_and_content_matching(
             "Host": "test_url2",
         },
         match_content=b"This is the body",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1865,9 +1801,7 @@ async def test_headers_not_matching_and_method_and_url_and_content_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_url_and_headers_not_matching_and_method_and_content_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1879,6 +1813,7 @@ async def test_url_and_headers_not_matching_and_method_and_content_matching(
             "Host": "test_url2",
         },
         match_content=b"This is the body",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1892,9 +1827,7 @@ async def test_url_and_headers_not_matching_and_method_and_content_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_method_and_url_and_headers_matching_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1906,6 +1839,7 @@ async def test_method_and_url_and_headers_matching_and_content_not_matching(
             "Host": "test_url",
         },
         match_content=b"This is the body2",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1919,9 +1853,7 @@ async def test_method_and_url_and_headers_matching_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_method_and_headers_matching_and_url_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1933,6 +1865,7 @@ async def test_method_and_headers_matching_and_url_and_content_not_matching(
             "Host": "test_url",
         },
         match_content=b"This is the body2",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1946,9 +1879,7 @@ async def test_method_and_headers_matching_and_url_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_method_and_url_matching_and_headers_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1960,6 +1891,7 @@ async def test_method_and_url_matching_and_headers_and_content_not_matching(
             "Host": "test_url2",
         },
         match_content=b"This is the body2",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -1973,9 +1905,7 @@ async def test_method_and_url_matching_and_headers_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_method_matching_and_url_and_headers_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -1987,6 +1917,7 @@ async def test_method_matching_and_url_and_headers_and_content_not_matching(
             "Host": "test_url2",
         },
         match_content=b"This is the body2",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -2000,9 +1931,7 @@ async def test_method_matching_and_url_and_headers_and_content_not_matching(
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_method_and_url_and_headers_and_content_not_matching(
     httpx_mock: HTTPXMock,
 ) -> None:
@@ -2014,6 +1943,7 @@ async def test_method_and_url_and_headers_and_content_not_matching(
             "Host": "test_url2",
         },
         match_content=b"This is the body2",
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -2149,11 +2079,10 @@ async def test_mutating_json(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 async def test_streams_are_not_cascading_resulting_in_maximum_recursion(
     httpx_mock: HTTPXMock,
 ) -> None:
-    httpx_mock.add_response(json={"abc": "def"})
+    httpx_mock.add_response(json={"abc": "def"}, is_reusable=True)
     async with httpx.AsyncClient() as client:
         tasks = [client.get("https://test_url") for _ in range(950)]
         await asyncio.gather(*tasks)
@@ -2269,9 +2198,7 @@ async def test_files_and_data_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_files_not_matching_name(httpx_mock: HTTPXMock, monkeypatch) -> None:
     # Ensure generated boundary will be fbe495efe4cd41b941ca13e254d6b018
     monkeypatch.setattr(
@@ -2280,7 +2207,7 @@ async def test_files_not_matching_name(httpx_mock: HTTPXMock, monkeypatch) -> No
         lambda length: b"\xfb\xe4\x95\xef\xe4\xcdA\xb9A\xca\x13\xe2T\xd6\xb0\x18",
     )
 
-    httpx_mock.add_response(match_files={"name2": ("file_name", b"File content")})
+    httpx_mock.add_response(match_files={"name2": ("file_name", b"File content")}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -2295,9 +2222,7 @@ async def test_files_not_matching_name(httpx_mock: HTTPXMock, monkeypatch) -> No
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_files_not_matching_file_name(httpx_mock: HTTPXMock, monkeypatch) -> None:
     # Ensure generated boundary will be fbe495efe4cd41b941ca13e254d6b018
     monkeypatch.setattr(
@@ -2306,7 +2231,7 @@ async def test_files_not_matching_file_name(httpx_mock: HTTPXMock, monkeypatch) 
         lambda length: b"\xfb\xe4\x95\xef\xe4\xcdA\xb9A\xca\x13\xe2T\xd6\xb0\x18",
     )
 
-    httpx_mock.add_response(match_files={"name": ("file_name2", b"File content")})
+    httpx_mock.add_response(match_files={"name": ("file_name2", b"File content")}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -2321,9 +2246,7 @@ async def test_files_not_matching_file_name(httpx_mock: HTTPXMock, monkeypatch) 
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_files_not_matching_content(httpx_mock: HTTPXMock, monkeypatch) -> None:
     # Ensure generated boundary will be fbe495efe4cd41b941ca13e254d6b018
     monkeypatch.setattr(
@@ -2332,7 +2255,7 @@ async def test_files_not_matching_content(httpx_mock: HTTPXMock, monkeypatch) ->
         lambda length: b"\xfb\xe4\x95\xef\xe4\xcdA\xb9A\xca\x13\xe2T\xd6\xb0\x18",
     )
 
-    httpx_mock.add_response(match_files={"name": ("file_name", b"File content2")})
+    httpx_mock.add_response(match_files={"name": ("file_name", b"File content2")}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -2347,9 +2270,7 @@ async def test_files_not_matching_content(httpx_mock: HTTPXMock, monkeypatch) ->
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_files_matching_but_data_not_matching(
     httpx_mock: HTTPXMock, monkeypatch
 ) -> None:
@@ -2363,6 +2284,7 @@ async def test_files_matching_but_data_not_matching(
     httpx_mock.add_response(
         match_files={"name": ("file_name", b"File content")},
         match_data={"field": "value"},
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -2391,12 +2313,11 @@ async def test_timeout_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_timeout_not_matching(httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(
-        match_extensions={"timeout": {"connect": 5, "read": 5, "write": 10, "pool": 5}}
+        match_extensions={"timeout": {"connect": 5, "read": 5, "write": 10, "pool": 5}},
+        is_optional=True,
     )
 
     async with httpx.AsyncClient() as client:
@@ -2421,11 +2342,9 @@ async def test_extensions_matching(httpx_mock: HTTPXMock) -> None:
 
 
 @pytest.mark.asyncio
-@pytest.mark.httpx_mock(
-    assert_all_responses_were_requested=False, assert_all_requests_were_expected=False
-)
+@pytest.mark.httpx_mock(assert_all_requests_were_expected=False)
 async def test_extensions_not_matching(httpx_mock: HTTPXMock) -> None:
-    httpx_mock.add_response(match_extensions={"test": "value"})
+    httpx_mock.add_response(match_extensions={"test": "value"}, is_optional=True)
 
     async with httpx.AsyncClient() as client:
         with pytest.raises(httpx.TimeoutException) as exception_info:
@@ -2435,3 +2354,60 @@ async def test_extensions_not_matching(httpx_mock: HTTPXMock) -> None:
             == """No response can be found for GET request on https://test_url with {'test': 'value2'} extensions amongst:
 - Match any request with {'test': 'value'} extensions"""
         )
+
+
+@pytest.mark.asyncio
+async def test_optional_response_not_matched(httpx_mock: HTTPXMock) -> None:
+    # This response is optional and the fact that it was never requested should not trigger anything
+    httpx_mock.add_response(url="https://test_url", is_optional=True)
+    httpx_mock.add_response(url="https://test_url2")
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("https://test_url2")
+    assert response.content == b""
+
+
+@pytest.mark.asyncio
+async def test_optional_response_matched(httpx_mock: HTTPXMock) -> None:
+    # This response is optional and the fact that it was never requested should not trigger anything
+    httpx_mock.add_response(url="https://test_url", is_optional=True)
+    httpx_mock.add_response(url="https://test_url2")
+
+    async with httpx.AsyncClient() as client:
+        response1 = await client.get("https://test_url")
+        response2 = await client.get("https://test_url2")
+    assert response1.content == b""
+    assert response2.content == b""
+
+
+@pytest.mark.asyncio
+@pytest.mark.httpx_mock(assert_all_responses_were_requested=False)
+async def test_mandatory_response_matched(httpx_mock: HTTPXMock) -> None:
+    # This response is optional and the fact that it was never requested should not trigger anything
+    httpx_mock.add_response(url="https://test_url")
+    # This response MUST be requested (overrides global settings via marker)
+    httpx_mock.add_response(url="https://test_url2", is_optional=False)
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("https://test_url2")
+    assert response.content == b""
+
+
+@pytest.mark.asyncio
+async def test_multi_response_matched_once(httpx_mock: HTTPXMock) -> None:
+    httpx_mock.add_response(url="https://test_url", is_reusable=True)
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get("https://test_url")
+    assert response.content == b""
+
+
+@pytest.mark.asyncio
+async def test_multi_response_matched_twice(httpx_mock: HTTPXMock) -> None:
+    httpx_mock.add_response(url="https://test_url", is_reusable=True)
+
+    async with httpx.AsyncClient() as client:
+        response1 = await client.get("https://test_url")
+        response2 = await client.get("https://test_url")
+    assert response1.content == b""
+    assert response2.content == b""